+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ### New features
 
+#### SCP-16 Add Firebase authentication
+  * Added basic Firebase authentication
+    * Added endpoints
+      * `POST: http://localhost:8000/signup/`
+      * `POST: http://localhost:8000/login/`
+      * `POST: http://localhost:8000/refresh-token/`
+
 #### SCP-15 Create simple Django API
   * Add basic Django rest API
   * Add changelog